3. MLE/3. Deploy model/services/requirements.txt
autofeat==2.1.3
catboost==1.2.7
category-encoders==2.6.4
cloudpickle==3.1.0
Fastapi==0.115.5
joblib==1.4.2
pandas==2.2.3
pydantic-settings==2.6.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.3.1
uvicorn==0.32.1
prometheus-fastapi-instrumentator==7.0.0
